LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/css/cssom-view/smooth-scroll-in-load-event-expected.txt


FAIL Smooth scroll in load event handler assert_equals: Should not set scrollLeft immediately expected 0 but got 800
